added tool Utils and moved PickaxeMineable.cs interaction to Pickaxe.cs
