+++ UnityProject/Assets/Scripts/Objects/Mining/PickaxeMineable.cs
@@ -1,42 +1,15 @@
-﻿using System.Collections;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Objects.Mining
 {
 	/// <summary>
 	/// Simple script to allow objects to be mined. Destroys the object upon completion.
 	/// </summary>
-	public class PickaxeMineable : MonoBehaviour, ICheckedInteractable<HandApply>
+	public class PickaxeMineable : MonoBehaviour
 	{
-		[SerializeField]
-		private float mineTime = 3;
+		[SerializeField] private float mineTime = 3;
+		public float MineTime => mineTime;
 
-		private string objectName;
-
-		public bool WillInteract(HandApply interaction, NetworkSide side)
-		{
-			objectName = gameObject.ExpensiveName();
-
-			if (DefaultWillInteract.Default(interaction, side) == false) return false;
-
-			return Validations.HasItemTrait(interaction.UsedObject, CommonTraits.Instance.Pickaxe);
-		}
-
-		public void ServerPerformInteraction(HandApply interaction)
-		{
-			var calculatedMineTime = interaction.HandObject.GetComponent<Pickaxe>().TimeMultiplier * mineTime;
-
-			ToolUtils.ServerUseToolWithActionMessages(
-					interaction, calculatedMineTime,
-					$"You start mining the {objectName}...",
-					$"{interaction.Performer.ExpensiveName()} starts mining the {objectName}...",
-					default, default,
-					() =>
-					{
-						SoundManager.PlayNetworkedAtPos(SingletonSOSounds.Instance.BreakStone,
-								interaction.PerformerPlayerScript.WorldPos, sourceObj: interaction.Performer);
-						Despawn.ServerSingle(gameObject);
-					});
-		}
+	
 	}
 }
